feat: move mermaid rendering to plugin
